+++ vite-helper/cdn-resolve.ts
@@ -6,11 +6,9 @@ import { defineResolve } from 'vite-plugin-cdn2/resolve'
  * 自定义 vite-plugin-cdn2 resolve 方法。
  * 默认使用 cdnjs 作为 cdn 服务商。
  *
- * @see https://cdnjs.cloudflare.com/ajax/libs/
+ * @see https://lib.baomitu.com/
  */
-export const cdnResolve = (
-  cdnBaseURL: string = 'https://cdnjs.cloudflare.com/ajax/libs/',
-) =>
+export const cdnResolve = (cdnBaseURL: string = 'https://lib.baomitu.com/') =>
   defineResolve({
     name: 'RayTemplateCdnResolve',
     setup({ extra }) {
@@ -20,7 +18,10 @@ export const cdnResolve = (
       return {
         url: url.href,
         injectTo: 'head-prepend',
-        attrs: {},
+        attrs: {
+          defer: true,
+          crossorigin: 'anonymous',
+        },
       }
     },
   })